// src/types/stats.ts
export type SoldBike = {
  id: number;
  model: string;
  manufacturer: string;
  primaryColor?: string;
  secondaryColor?: string;
  place: string;
  price: number;
  salePrice: number;
  discount: number;
  discountPercent: number;
  saleDate: string;
};
// export type Series<T> = {
//   label: string;
//   data: T[];
// };
export type PieChart = {
  id: number;
  name: string;
  quantity: number;
  value: number;
};

export type DateAndPrice = {
  date: string;
  price: number;
};
export type BarChartData<T extends string> = {
  [key in T]: string;
} & BarChartFields;

type BarChartFields = {
  [key: string]: number;
};